# docs/swagger.yaml
basePath: /im
definitions:
  enums.FriendReqStatus:
    enum:
    - pending
    - accepted
    - rejected
    type: string
    x-enum-varnames:
    - FriendReqStatusPending
    - FriendReqStatusAccepted
    - FriendReqStatusRejected
  enums.FriendStatus:
    enum:
    - active
    - blocked
    type: string
    x-enum-varnames:
    - FriendStatusActive
    - FriendStatusBlocked
...
  req.ExampleCreate:
    properties:
      description:
        description: 創建範例描述
        example: 一個武林高手
        type: string
      name:
        description: 創建範例名
        example: 小明
        type: string
    type: object
  req.ExampleDelete:
    properties:
      id:
        description: 刪除範例ID
        example: "1"
        type: string
    type: object
  req.ExampleGet:
    properties:
      description:
        description: 範例描述
        example: 取得描述
        type: string
      name:
        description: 範例名
        example: 名字
        type: string
    type: object
  req.ExampleGetList:
    properties:
      id:
        description: 範例ID列表
        example:
        - "1"
        - "2"
        - "3"
        - "4"
        - "5"
        - "6"
        items:
          type: string
        type: array
      index:
        description: 頁碼
        type: integer
      name:
        description: 範例名
        example: 名字
        type: string
      order:
        description: 排序
        example: id asc
        type: string
      size:
        description: 筆數
        type: integer
    required:
    - id
    type: object
  req.ExampleUpdate:
    properties:
      description:
        description: 修改範例描述
        example: 繼承小明武功但沒天分的孩子
        type: string
      id:
        description: 修改範例ID
        example: "1"
        type: string
      name:
        description: 修改範例名
        example: 小明的孩子
        type: string
    required:
    - id
    type: object
  req.FriendDelete:
    properties:
      id:
        type: string
    type: object
  req.FriendGetList:
    properties:
      index:
        description: 頁碼
        type: integer
      order:
        description: 排序
        example: id asc
        type: string
      size:
        description: 筆數
        type: integer
    type: object
  req.FriendMutualGet:
    properties:
      index:
        description: 頁碼
        type: integer
      order:
        description: 排序
        example: id asc
        type: string
      size:
        description: 筆數
        type: integer
      t_user_id:
        description: 對象用户ID
        type: string
    required:
    - t_user_id
    type: object
  req.FriendRequestsCreate:
    properties:
      user_name:
        description: 對象用戶username
        example: user
        type: string
    required:
    - user_name
    type: object
  req.FriendRequestsGetList:
    properties:
      index:
        description: 頁碼
        type: integer
      is_sender:
        description: 'true: 請求列表 false: 被請求列表'
        example: false
        type: boolean
      order:
        description: 排序
        example: id asc
        type: string
      size:
        description: 筆數
        type: integer
    type: object
  req.FriendRequestsUpdate:
    properties:
      id:
        description: 請求單ID
        type: string
      request_status:
        allOf:
        - $ref: '#/definitions/enums.FriendReqStatus'
        enum:
        - accepted
        - rejected
    required:
    - request_status
    type: object
  req.FriendUpdate:
    properties:
      f_user_id:
        description: 好友的用户ID
        type: string
      status:
        $ref: '#/definitions/enums.FriendStatus'
    required:
    - f_user_id
    type: object
  req.GroupInvitationCreate:
    properties:
      groupId:
        description: 群組ID
        type: string
      inviteeId:
        description: 被邀請者ID
        type: string
    type: object
  req.GroupInvitationDelete:
    properties:
      id:
        type: string
    type: object
  req.GroupInvitationGet:
    properties:
      groupID:
        description: 群组的唯一标识符
        type: string
      id:
        description: 此邀請標示
        type: string
      invitationStatus:
        description: 邀请的状态
        type: string
      inviteeID:
        description: 被邀请者的唯一标识符
        type: string
      inviterID:
        description: 邀请者的唯一标识符
        type: string
    type: object
  req.GroupInvitationGetList:
    properties:
      index:
        description: 頁碼
        type: integer
      order:
        description: 排序
        example: id asc
        type: string
      size:
        description: 筆數
        type: integer
    type: object
  req.GroupInvitationUpdate:
    type: object
...
  resp.APIResponse-resp_ExampleGet:
    properties:
      code:
        description: 回傳代碼
        type: string
      data:
        allOf:
        - $ref: '#/definitions/resp.ExampleGet'
        description: 資料
      msg:
        description: 訊息
        type: string
    type: object
  resp.APIResponse-resp_ExampleGetList:
    properties:
      code:
        description: 回傳代碼
        type: string
      data:
        allOf:
        - $ref: '#/definitions/resp.ExampleGetList'
        description: 資料
      msg:
        description: 訊息
        type: string
    type: object
  resp.APIResponse-resp_FriendGetList:
    properties:
      code:
        description: 回傳代碼
        type: string
      data:
        allOf:
        - $ref: '#/definitions/resp.FriendGetList'
        description: 資料
      msg:
        description: 訊息
        type: string
    type: object
  resp.APIResponse-resp_FriendRequestsGetList:
    properties:
      code:
        description: 回傳代碼
        type: string
      data:
        allOf:
        - $ref: '#/definitions/resp.FriendRequestsGetList'
        description: 資料
      msg:
        description: 訊息
        type: string
    type: object
  resp.APIResponse-resp_GroupInvitationGet:
    properties:
      code:
        description: 回傳代碼
        type: string
      data:
        allOf:
        - $ref: '#/definitions/resp.GroupInvitationGet'
        description: 資料
      msg:
        description: 訊息
        type: string
    type: object
  resp.APIResponse-resp_GroupInvitationGetList:
    properties:
      code:
        description: 回傳代碼
        type: string
      data:
        allOf:
        - $ref: '#/definitions/resp.GroupInvitationGetList'
        description: 資料
      msg:
        description: 訊息
        type: string
    type: object
...
  resp.APIResponse-string:
    properties:
      code:
        description: 回傳代碼
        type: string
      data:
        description: 資料
        type: string
      msg:
        description: 訊息
        type: string
    type: object
  resp.ExampleGet:
    properties:
      description:
        description: 描述
        type: string
      id:
        description: 數據ID
        type: string
      name:
        description: 範例名
        type: string
    type: object
  resp.ExampleGetList:
    properties:
      data:
        items:
          $ref: '#/definitions/resp.ExampleGet'
        type: array
      page:
        $ref: '#/definitions/resp.PageResponse'
    type: object
  resp.FriendGet:
    properties:
      f_user_id:
        type: string
      f_user_name:
        type: string
      id:
        type: string
      mute:
        type: boolean
      p_user_id:
        type: string
      p_user_name:
        type: string
      status:
        $ref: '#/definitions/enums.FriendStatus'
    type: object
  resp.FriendGetList:
    properties:
      data:
        items:
          $ref: '#/definitions/resp.FriendGet'
        type: array
      page:
        $ref: '#/definitions/resp.PageResponse'
    type: object
  resp.FriendRequestsGet:
    properties:
      createdAt:
        type: string
      id:
        type: string
      receiverID:
        type: string
      receiverName:
        type: string
      requestStatus:
        $ref: '#/definitions/enums.FriendReqStatus'
      senderID:
        type: string
      senderName:
        type: string
    type: object
  resp.FriendRequestsGetList:
    properties:
      data:
        items:
          $ref: '#/definitions/resp.FriendRequestsGet'
        type: array
      page:
        $ref: '#/definitions/resp.PageResponse'
    type: object
  resp.GroupInvitationGet:
    type: object
  resp.GroupInvitationGetList:
    properties:
      data:
        items:
          $ref: '#/definitions/resp.GroupInvitationGet'
        type: array
      page:
        $ref: '#/definitions/resp.PageResponse'
    type: object
...
  resp.PageResponse:
    properties:
      index:
        description: 頁碼
        type: integer
      pages:
        description: 總頁數
        type: integer
      size:
        description: 筆數
        type: integer
      total:
        description: 總筆數
        type: integer
    type: object
...
host: localhost:9000
info:
  contact:
    name: API Support
  description: This is a project im.
  termsOfService: http://swagger.io/terms/
  title: Im
  version: "1.0"
paths:
...
  /example:
    delete:
      parameters:
      - description: param
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/req.ExampleDelete'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.APIResponse-string'
      summary: Delete
      tags:
      - example
    get:
      parameters:
      - description: param
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/req.ExampleGetList'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.APIResponse-resp_ExampleGetList'
      summary: GetList
      tags:
      - example
    post:
      parameters:
      - description: param
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/req.ExampleCreate'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.APIResponse-string'
      summary: Create
      tags:
      - example
    put:
      parameters:
      - description: param
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/req.ExampleUpdate'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.APIResponse-string'
      summary: Update
      tags:
      - example
  /example/:id:
    get:
      parameters:
      - description: param
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/req.ExampleGet'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.APIResponse-resp_ExampleGet'
      summary: Get
      tags:
      - example
  /friend:
    delete:
      parameters:
      - description: param
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/req.FriendDelete'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.APIResponse-string'
      summary: 刪除與指定用戶的好友關係
      tags:
      - friend
    get:
      parameters:
      - description: param
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/req.FriendGetList'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.APIResponse-resp_FriendGetList'
      summary: 獲取用戶的好友列表
      tags:
      - friend
    put:
      parameters:
      - description: param
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/req.FriendUpdate'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.APIResponse-string'
      summary: 更新與指定用戶的好友關係（接受/拒絕/阻止）
      tags:
      - friend
  /friend-requests:
    get:
      parameters:
      - description: param
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/req.FriendRequestsGetList'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.APIResponse-resp_FriendRequestsGetList'
      summary: 好友請求列表
      tags:
      - FriendRequests
    post:
      parameters:
      - description: param
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/req.FriendRequestsCreate'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.APIResponse-string'
        "400":
          description: 无效的用户
          schema:
            $ref: '#/definitions/resp.APIResponse-string'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/resp.APIResponse-string'
        "409":
          description: Friend request already exists between these users.
          schema:
            $ref: '#/definitions/resp.APIResponse-string'
        "500":
          description: 未知错误
          schema:
            $ref: '#/definitions/resp.APIResponse-string'
      summary: 向指定用戶ID發送好友請求
      tags:
      - FriendRequests
    put:
      parameters:
      - description: param
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/req.FriendRequestsUpdate'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.APIResponse-string'
        "400":
          description: 无效的ID
          schema:
            $ref: '#/definitions/resp.APIResponse-string'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/resp.APIResponse-string'
        "500":
          description: 未知错误
          schema:
            $ref: '#/definitions/resp.APIResponse-string'
      summary: 接受或拒絕來自id的好友請求
      tags:
      - FriendRequests
  /friend/blocked:
    get:
      parameters:
      - description: param
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/req.FriendGetList'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.APIResponse-string'
      summary: 獲取指定用戶ID的已封鎖好友列表
      tags:
      - friend
  /friend/mutual:
    get:
      parameters:
      - description: param
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/req.FriendMutualGet'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.APIResponse-string'
      summary: 獲取指定用戶ID與另一指定用戶ID的共同好友列表
      tags:
      - friend
...
  /group-invitation:
    delete:
      parameters:
      - description: param
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/req.GroupInvitationDelete'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.APIResponse-string'
      summary: Delete
      tags:
      - groupInvitation
    get:
      parameters:
      - description: param
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/req.GroupInvitationGetList'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.APIResponse-resp_GroupInvitationGetList'
      summary: GetList
      tags:
      - groupInvitation
    post:
      parameters:
      - description: param
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/req.GroupInvitationCreate'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.APIResponse-string'
      summary: 新建邀請碼
      tags:
      - groupInvitation
    put:
      parameters:
      - description: param
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/req.GroupInvitationUpdate'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.APIResponse-string'
      summary: Update
      tags:
      - groupInvitation
  /group-invitation/:id:
    get:
      parameters:
      - description: param
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/req.GroupInvitationGet'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.APIResponse-resp_GroupInvitationGet'
      summary: Get
      tags:
      - groupInvitation
...
schemes:
- http
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: token
    type: apiKey
swagger: "2.0"